Got rid of AbstractTxTask
